# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Documentation API",
        "description": "API untuk integrasi dengan aplikasi Workshop. Endpoint ini dipakai aplikasi terpisah untuk menerima Permintaan Unit yang sudah diteruskan, menyetujui penyiapan unit, dan mengupdate kesiapan tiap item.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "SSB Project API Server"
        }
    ],
    "paths": {
        "/unit-replacements": {
            "get": {
                "tags": [
                    "Unit Replacement"
                ],
                "summary": "List Penggantian Unit (PTU) yang sudah diteruskan ke Workshop",
                "description": "Default-nya hanya menampilkan status FORWARDED_TO_WORKSHOP. Gunakan parameter ?status=APPROVED_FROM_WORKSHOP, ?status=REJECTED_FROM_WORKSHOP, kombinasi (comma-separated), atau ?status=all untuk mengubah filter.",
                "operationId": "962795c9fdf453da98b798378b8536fa",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status. Default: FORWARDED_TO_WORKSHOP. Boleh comma-separated atau 'all'.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "FORWARDED_TO_WORKSHOP"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman (1-100). Default: 25.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List PTU (paginated)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Parameter status tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/unit-replacements/{uid}": {
            "get": {
                "tags": [
                    "Unit Replacement"
                ],
                "summary": "Detail satu PTU",
                "operationId": "a0312ead7ffab271916860577042ecc5",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "UID (UUID) PTU",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail PTU"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/unit-replacements/{uid}/status": {
            "patch": {
                "tags": [
                    "Unit Replacement"
                ],
                "summary": "Update status PTU",
                "description": "Workshop memakai endpoint ini untuk menyetujui atau menolak PTU yang diteruskan. Hanya transisi yang diizinkan UnitReplacementStatus::canTransitionTo() yang diterima.",
                "operationId": "828b2c80e9120e2855614afb0a7e4b13",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "UID (UUID) PTU",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Status tujuan. Untuk Workshop: APPROVED_FROM_WORKSHOP atau REJECTED_FROM_WORKSHOP.",
                                        "type": "string",
                                        "enum": [
                                            "DRAFT",
                                            "SUBMITTED",
                                            "APPROVED",
                                            "REJECTED",
                                            "FORWARDED_TO_WORKSHOP",
                                            "APPROVED_FROM_WORKSHOP",
                                            "REJECTED_FROM_WORKSHOP"
                                        ],
                                        "example": "APPROVED_FROM_WORKSHOP"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Unit pengganti siap kirim",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diupdate"
                    },
                    "422": {
                        "description": "Status tidak valid atau transisi tidak diizinkan"
                    },
                    "404": {
                        "description": "PTU tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}